pyproject.toml
[project]
name = "bricksops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "databricks-connect>=16.0.0",
    "mypy>=1.14.0",
    "pyspark>=3.5.4",
    "pytest>=8.3.4",
    "requests-mock>=1.12.1",
    "ruff>=0.8.4",
    "types-pyyaml>=6.0.12.20241221",
    "types-requests>=2.32.0.20241016",
]

[tool.pytest.ini_options]
testpaths = "tests"
pythonpath = "."
